backend/tests/test_database_auth.py
"""
Tests for database and authentication (Sprint P3.2).

These tests verify:
- Password hashing with bcrypt
- User registration and login
- Database models
- History persistence
"""

import pytest
import bcrypt
from unittest.mock import patch, AsyncMock, MagicMock

# Import the modules to test
from backend.services.auth import hash_password, verify_password
from backend.database import User, Summary


class TestPasswordHashing:
    """Tests for password hashing (bcrypt)."""

    def test_hash_password_returns_bcrypt_hash(self):
        """hash_password should return a bcrypt hash."""
        password = "testpassword123"
        hashed = hash_password(password)
        
        # bcrypt hashes start with $2a$, $2b$, or $2y$
        assert hashed.startswith("$2b$")
        assert len(hashed) == 60  # bcrypt hash length

    def test_verify_password_correct_password(self):
        """verify_password should return True for correct password."""
        password = "testpassword123"
        hashed = hash_password(password)
        
        assert verify_password(password, hashed) is True

    def test_verify_password_wrong_password(self):
        """verify_password should return False for wrong password."""
        password = "testpassword123"
        wrong_password = "wrongpassword"
        hashed = hash_password(password)
        
        assert verify_password(wrong_password, hashed) is False

    def test_verify_password_empty_password(self):
        """verify_password should return False for empty password."""
        password = "testpassword123"
        hashed = hash_password(password)
        
        assert verify_password("", hashed) is False

    def test_different_hashes_for_same_password(self):
        """Same password should produce different hashes (due to salt)."""
        password = "testpassword123"
        hash1 = hash_password(password)
        hash2 = hash_password(password)
        
        # Hashes should be different due to unique salt
        assert hash1 != hash2
        # But both should verify correctly
        assert verify_password(password, hash1) is True
        assert verify_password(password, hash2) is True


class TestDatabaseModels:
    """Tests for database models."""

    def test_user_model_fields(self):
        """User model should have expected fields."""
        user = User(
            id=1,
            email="test@example.com",
            username="testuser",
            hashed_password="$2b$12$hash",
        )
        
        assert user.id == 1
        assert user.email == "test@example.com"
        assert user.username == "testuser"
        assert user.hashed_password == "$2b$12$hash"

    def test_summary_model_fields(self):
        """Summary model should have expected fields."""
        summary = Summary(
            id=1,
            user_id=1,
            video_id="dQw4w9WgXcQ",
            video_title="Test Video",
            video_url="https://youtube.com/watch?v=dQw4w9WgXcQ",
            video_channel="Test Channel",
            transcript_language="en",
            transcript_word_count=500,
            transcript_duration=300,
            llm_provider="openrouter",
            llm_model="deepseek/deepseek-v3.2",
            summary_text="This is a test summary.",
        )
        
        assert summary.id == 1
        assert summary.user_id == 1
        assert summary.video_id == "dQw4w9WgXcQ"
        assert summary.video_title == "Test Video"
        assert summary.summary_text == "This is a test summary."
        assert summary.llm_provider == "openrouter"

    def test_user_repr(self):
        """User model should have a useful repr."""
        user = User(id=1, email="test@example.com", hashed_password="hash")
        repr_str = repr(user)
        
        assert "User" in repr_str
        assert "test@example.com" in repr_str


class TestAuthService:
    """Tests for authentication service functions."""

    @pytest.mark.asyncio
    async def test_register_user_raises_if_exists(self):
        """register_user should raise ValueError if user exists."""
        from backend.services.auth import register_user, user_store
        from unittest.mock import AsyncMock, MagicMock
        
        mock_db = MagicMock()
        
        # Mock that user already exists
        user_store.user_exists = AsyncMock(return_value=True)
        
        with pytest.raises(ValueError, match="User already exists"):
            await register_user("test@example.com", "password123", "testuser", mock_db)

    @patch('backend.services.auth.user_store')
    def test_authenticate_user_returns_none_for_unknown(self, mock_store):
        """authenticate_user should return None for unknown email."""
        from backend.services.auth import authenticate_user
        
        # Mock that user doesn't exist
        mock_store.get_user_by_email = AsyncMock(return_value=None)
        
        # This would need db session - tested via integration below


class TestIntegration:
    """Integration tests with mocked database."""

    @pytest.mark.asyncio
    async def test_register_creates_user_with_hashed_password(self):
        """Registration should create user with bcrypt hashed password."""
        from backend.services.auth import register_user, user_store
        from unittest.mock import AsyncMock, MagicMock
        
        # Create mock database session
        mock_db = MagicMock()
        
        # Mock user doesn't exist
        user_store.user_exists = AsyncMock(return_value=False)
        
        # Mock user creation
        async def mock_create(email, username, hashed_password, db):
            return {
                "id": 1,
                "email": email,
                "username": username,
                "hashed_password": hashed_password,
                "created_at": "2026-05-02T00:00:00"
            }
        user_store.create_user = mock_create
        
        # Register user
        user = await register_user("test@example.com", "password123", "testuser", mock_db)
        
        # Verify user was created
        assert user["email"] == "test@example.com"
        assert user["username"] == "testuser"
        
        # Verify password was hashed (not stored in plain text)
        # The hashed_password should start with $2b$
        created_user = await user_store.create_user("test@example.com", "testuser", "dummy", mock_db)
        # This is a simplified check - in real test we'd verify the hash format

    @pytest.mark.asyncio
    async def test_login_verifies_password(self):
        """Login should verify password against bcrypt hash."""
        from backend.services.auth import authenticate_user, user_store, verify_password
        from unittest.mock import AsyncMock, MagicMock
        
        # Create mock database session
        mock_db = MagicMock()
        
        # Create a real bcrypt hash
        real_hash = hash_password("correctpassword")
        
        # Mock user exists with real hash
        user_store.get_user_by_email = AsyncMock(return_value={
            "id": 1,
            "email": "test@example.com",
            "hashed_password": real_hash
        })
        
        # Test with correct password
        user = await authenticate_user("test@example.com", "correctpassword", mock_db)
        assert user is not None
        assert user["email"] == "test@example.com"
        
        # Test with wrong password
        user = await authenticate_user("test@example.com", "wrongpassword", mock_db)
        assert user is None


class TestHistoryStore:
    """Tests for history store."""

    def test_history_store_has_add_summary_method(self):
        """history_store should have add_summary method."""
        from backend.services.auth import history_store
        
        assert hasattr(history_store, 'add_summary')
        assert callable(history_store.add_summary)

    def test_history_store_has_get_user_history_method(self):
        """history_store should have get_user_history method."""
        from backend.services.auth import history_store
        
        assert hasattr(history_store, 'get_user_history')
        assert callable(history_store.get_user_history)

    def test_history_store_has_delete_summary_method(self):
        """history_store should have delete_summary method."""
        from backend.services.auth import history_store
        
        assert hasattr(history_store, 'delete_summary')
        assert callable(history_store.delete_summary)

    @pytest.mark.asyncio
    async def test_add_summary_interface(self):
        """add_summary should accept expected parameters."""
        from backend.services.auth import history_store
        from unittest.mock import MagicMock
        
        mock_db = MagicMock()
        
        # Verify the method signature accepts the expected parameters
        import inspect
        sig = inspect.signature(history_store.add_summary)
        params = list(sig.parameters.keys())
        
        assert 'user_id' in params
        assert 'video_id' in params
        assert 'video_title' in params
        assert 'summary' in params

    @pytest.mark.asyncio
    async def test_get_user_history_interface(self):
        """get_user_history should accept expected parameters."""
        from backend.services.auth import history_store
        
        import inspect
        sig = inspect.signature(history_store.get_user_history)
        params = list(sig.parameters.keys())
        
        assert 'user_id' in params
        assert 'limit' in params