requirements.txt
youtube-transcript-api
yt-dlp
tiktoken
python-dotenv

# LLM provider SDK (OpenRouter)
openai

# FastAPI stack (used in later phases; harmless for Phase 1 environment readiness)
fastapi
uvicorn[standard]

# Authentication (TSK-0301)
python-jose[cryptography]
passlib[bcrypt]
httpx

# Database (Sprint P3.2 - MySQL)
sqlalchemy
aiomysql
pymysql
